fix context/client bug
Previously, it was impossible to instantiate multiple Client instances,
as they'd all end up using the same handler. This is because the
`on_message_complete` callback was static, and would only every
reference a single handler.

This has been fixed by adding an `instance` property to `Context` which
holds the relevant `Server` or `Client` instance. This is used to avoid
having the static handler lambda.

In addition, most of the lambdas internal to the library have been
removed for simplicity. This eliminates the need for things like wrappers
around lambdas defined with `[&]`, which can have confusing results.